refactor: extract windows, surfaces, and onboarding into AppDelegate+WindowsAndSurfaces.swift
Move all remaining non-core methods from AppDelegate.swift into a new
extension file, reducing AppDelegate.swift to ~322 lines of pure core
orchestration (property declarations, applicationDidFinishLaunching,
proceedToApp, applicationWillTerminate).

Methods moved (~525 lines):
- Surface wiring: setupSurfaceManager, setupToolConfirmationNotifications,
  setupSecretPromptManager
- Window management: setupWindowObserver, scheduleActivationPolicyRevert,
  applicationShouldHandleReopen, ensureMainWindowExists, showMainWindow,
  observeConversationZoomEnabled, observeAssistantStatus,
  observeConversationBadge, formatDockConversationBadge,
  applyDockConversationBadge, refreshDockConversationBadge
- Onboarding: replayOnboarding, showOnboarding, lockfileHasAssistants,
  isGatewayHealthy, removeLockfileEntry, wakeUpGreeting
- Settings: showAboutPanel, showSettingsWindow, showSettingsTab

Access widened from private to internal where needed for cross-file
extension access. Pure mechanical move with zero logic changes.

Part of plan: appdelegate-refactor.md (PR 7 of 8)
